Replace  auth
