+++ src/components/app/index.js
@@ -2,10 +2,10 @@ import styles from './app.module.css';
 import AppHeader from '../app-header/app-header';
 import { Routes, Route, useLocation, useNavigate } from 'react-router-dom';
 import Home from '../../pages/home';
-import Login from '../../pages/login';
-import Register from '../../pages/register';
-import ForgotPassword from '../../pages/forgot-password';
-import ResetPassword from '../../pages/reset-password';
+import Login from '../../pages/auth/login';
+import Register from '../../pages/auth/register';
+import ForgotPassword from '../../pages/auth/forgot-password';
+import ResetPassword from '../../pages/auth/reset-password';
 import Profile from '../../pages/profile';
 import ProfileForm from '../../pages/profile/profile-form';
 import OrderHistory from '../../pages/profile/order-history';
