+++ src/pages/auth/forgot-password.js
@@ -1,7 +1,7 @@
 import { EmailInput } from '@ya.praktikum/react-developer-burger-ui-components';
 import { Link } from 'react-router-dom';
-import AuthForm from '../components/auth/auth-form';
-import styles from '../components/auth/auth-form.module.css';
+import AuthForm from '../../components/auth/auth-form';
+import styles from '../../components/auth/auth-form.module.css';
 import { useState } from 'react';
 
 const ForgotPassword = () => {
